+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Storage/LedgerDB/V1/Flush.hs
@@ -4,8 +4,8 @@ module Ouroboros.Consensus.Storage.LedgerDB.V1.Flush (
   ) where
 
 import           Ouroboros.Consensus.Ledger.Abstract
-import           Ouroboros.Consensus.Storage.LedgerDB.BackingStore
-import           Ouroboros.Consensus.Storage.LedgerDB.DbChangelog
+import           Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore
+import           Ouroboros.Consensus.Storage.LedgerDB.V1.DbChangelog
 import           Ouroboros.Consensus.Util.IOLike
 
 flushLedgerDB :: (MonadSTM m, GetTip l, HasLedgerTables l)
